requirements.txt
SQLAlchemy~=2.0.32
geopy~=2.4.1
aiogram~=3.12.0
requests~=2.32.3
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
asyncpg~=0.29.0
lxml~=5.3.0